requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitsandbytes==0.42.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datasets==3.3.2
defusedxml==0.7.1
dill==0.3.8
filelock==3.18.0
fonttools==4.57.0
fpdf2==2.4.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
googletrans==4.0.2
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
narwhals==1.32.0
networkx==3.4.2
numexpr==2.10.1
numpy==1.24.0
ollama==0.4.7
optimum==1.24.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==25.0
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.21
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
